+++ src/font/FontFamily.native.ts
@@ -1,5 +1,5 @@
-import * as Font                                                             from "expo-font";
-import {FONT_TYPES, FONT_WEIGHTS, FontFamily, FontFamilyConfig, FontVariant} from "./Fonts";
+import * as Font                                                                          from "expo-font";
+import {FONT_TYPES, FONT_WEIGHT_VALUE_TO_NAME, FontFamily, FontFamilyConfig, FontVariant} from "./Fonts";
 
 export const fontFamilyMap: Record<string, FontFamily>   = {};
 export const fontVariantMap: Record<string, FontVariant> = {};
@@ -42,7 +42,7 @@ export const createFontFamily = (
   const fontFamily: any = Object.assign(() => fontFamily.regular(), {
     weight: (weight: string) => {
       if (weight.endsWith("00")) {
-        return fontFamily[FONT_WEIGHTS[parseInt(weight)]]();
+        return fontFamily[FONT_WEIGHT_VALUE_TO_NAME[parseInt(weight)]]();
       }
       return fontFamily[weight]();
     },
@@ -55,16 +55,17 @@ export const createFontFamily = (
     const fontName   = `${name}__${type}`;
     fontFamily[type] = () => {
       if (!isFontLoading(fontName) && !isFontLoaded(fontName)) {
-        const resource = config[type] as string;
-        if (resource) {
+        const resource = config[type];
+        const location = typeof resource === "string" ? resource : (resource as any)[Object.keys(resource as any)[0] ];
+        if (location) {
           fontVariantMap[fontName] = {
-            weight    : parseInt(Object.keys(FONT_WEIGHTS).find(weight => FONT_WEIGHTS[weight as any] === type) as string),
+            weight    : parseInt(Object.keys(FONT_WEIGHT_VALUE_TO_NAME).find(weight => FONT_WEIGHT_VALUE_TO_NAME[weight as any] === type) as string),
             fontFamily: fontFamily,
-            resource  : resource,
+            resource  : location,
           };
 
           Font.loadAsync({
-            [fontName]: resource,
+            [fontName]: location,
           }).then(() => {
             getFontCallbackList(fontName).forEach(callback => callback());
           });