+++ src/font/FontFamily.ts
@@ -1,6 +1,15 @@
-import {FONT_TYPES, FONT_WEIGHTS, FontFamily, FontFamilyConfig, FontFamilyOptions, FontVariant} from "./Fonts";
-import {StyleEnvironment}                                                                       from "../index";
-import {isWeb}                                                                                  from "../Utils";
+import {
+  FONT_FORMAT_NAMES,
+  FONT_TYPES,
+  FONT_WEIGHT_NAME_TO_VALUE,
+  FONT_WEIGHT_VALUE_TO_NAME,
+  FontFamily,
+  FontFamilyConfig,
+  FontFamilyOptions,
+  FontVariant,
+  FontWeightConfig,
+}                         from "./Fonts";
+import {StyleEnvironment} from "../index";
 
 export const fontFamilyMap: Record<string, FontFamily>   = {};
 export const fontVariantMap: Record<string, FontVariant> = {};
@@ -38,7 +47,7 @@ export const createFontFamily = (
   const fontFamily: any = Object.assign(() => fontFamily.regular(), {
     weight: (weight: string) => {
       if (weight.endsWith("00")) {
-        return fontFamily[FONT_WEIGHTS[parseInt(weight)]]();
+        return fontFamily[FONT_WEIGHT_VALUE_TO_NAME[parseInt(weight)]]();
       }
       return fontFamily[weight]();
     },
@@ -51,22 +60,30 @@ export const createFontFamily = (
 
   for (const type of FONT_TYPES) {
     const fontName   = `${name}__${type}`;
-    const styleValue = isWeb() ? [fontName, ...options?.fallbacks || []].join(",") : fontName;
+    const styleValue = [name, ...options?.fallbacks || []].join(",");
 
     fontFamily[type] = () => {
       if (!isFontLoading(fontName)) {
-        const resource = config[type] as string;
+        const resource = config[type] as FontWeightConfig;
         if (resource) {
-          const fileType = resource.split(".")[1];
-          const css      = `@font-face{font-family: '${fontName}';font-display: swap;src: url(${resource})}`;
-          StyleEnvironment.updateHeadElement(`font-family-style(${fontName})`, "style", {"data-font-family": fontName}, css);
-          StyleEnvironment.updateHeadElement(`font-family-preload(${fontName})`, "link", {
-            "rel"        : "preload",
-            "href"       : resource,
-            "as"         : "font",
-            "type"       : `font/${fileType}`,
-            "crossorigin": "anonymous",
-          });
+          const files: { type: string; location: string }[] = [];
+          if (typeof resource === "string") {
+            const fileType = resource.split(".")[1];
+            files.push({type: fileType, location: resource});
+          } else {
+            for (const key of ["woff2", "woff", "ttf", "eot"]) {
+              const value = (resource as any)[key];
+              if (value) {
+                files.push({type: key, location: value});
+              }
+            }
+          }
+
+          const css = `@font-face{font-family: '${name}';font-style: ${type.includes("Italic") ? "italic" : "normal"};font-display: swap;font-weight:${FONT_WEIGHT_NAME_TO_VALUE[type]};src: local('${name}'),${files.map(file => `url('${file.location}') format('${FONT_FORMAT_NAMES[file.type]}')`).join(",")};}`;
+          StyleEnvironment.updateHeadElement(`font-family-style(${fontName})`, "style", {
+            "data-font-family": name,
+            "data-font-weight": type,
+          }, css);
           loadingMap[fontName] = true;
         }
       }