+++ src/rule/MediaRule.ts
@@ -10,7 +10,8 @@ export interface MediaQuery {
 }
 
 const media = createStyleRuleType<MediaQuery>("media", {
-  check(options: MediaQuery) {
+  check(instance) {
+    const {options} = instance;
     const screenWidth  = Math.round(StyleEnvironment.getScreenWidth());
     const screenHeight = Math.round(StyleEnvironment.getScreenHeight());