+++ src/rule/RtlRule.ts
@@ -2,7 +2,7 @@ import {createStyleRuleType} from "./StyleRule";
 import {I18nManager}         from "react-native";
 
 const rtl = createStyleRuleType<void>("rtl", {
-  check(options, session) {
+  check() {
     return I18nManager.isRTL;
   },
 });