+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Features / Changes
 
+- massive internal re-write to prep for rules within rules
 - SSR support
 - add `FontFamily` fallbacks for web
 - add `child()` rule